# internal/api/server.go
package api

import (
	"context"
	"fmt"
	"log/slog"
	"net/http"
	"os"
	"strings"
	"time"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	// This import is needed for swagger side-effects, it will be generated by swag init
	_ "github.com/mario-ezquerro/gubernator/docs"
	"github.com/mario-ezquerro/gubernator/internal/aqueducts"
	"github.com/mario-ezquerro/gubernator/internal/caddy"
	"github.com/mario-ezquerro/gubernator/internal/coredns"
	"github.com/mario-ezquerro/gubernator/internal/db"
	"github.com/mario-ezquerro/gubernator/internal/monitor"
	"github.com/mario-ezquerro/gubernator/internal/telemetry"
	"github.com/mario-ezquerro/gubernator/internal/web"
)

// dbPath returns the path to the SQLite database file.
// It respects the GBNT_DATA_DIR env var so Docker volumes work seamlessly.
func dbPath() string {
	dir := os.Getenv("GBNT_DATA_DIR")
	if dir == "" {
		dir = "/data"
	}
	// If running outside Docker (no /data dir), fall back to current dir
	if _, err := os.Stat(dir); os.IsNotExist(err) {
		dir = "."
	}
	return dir + "/gubernator.db"
}

// Start initializes the Gin router and starts the REST API server on port 4000.
// It blocks until ctx is cancelled, then shuts down gracefully.
func Start(ctx context.Context) error {
	if err := db.Init(dbPath()); err != nil {
		return fmt.Errorf("database init: %w", err)
	}

	// ── CoreDNS: Ensure gbnt-net network and CoreDNS container are running ──
	if err := coredns.EnsureNetwork(); err != nil {
		slog.Warn("CoreDNS: failed to create gbnt-net network", "err", err)
	} else {
		go func() {
			// Small delay to allow Docker daemon to fully register the network
			time.Sleep(2 * time.Second)
			if err := coredns.EnsureRunning(); err != nil {
				slog.Warn("CoreDNS: failed to start gbnt-coredns container", "err", err)
			}
			if err := caddy.EnsureRunning(); err != nil {
				slog.Warn("Caddy: failed to start gbnt-caddy container", "err", err)
			}
			if err := coredns.RegisterInDB(db.GetDB()); err != nil {
				slog.Warn("core stack: failed to register in database", "err", err)
			}
			aqueducts.GenerateHostsFile()
		}()
	}

	go startWatchtowers(ctx)
	go web.StartDashboard()
	go startTelemetryServer(ctx)
	go startLocalExecutor(ctx)

	// Auto-deploy SRE Monitoring Stack if GBNT_MONITOR=true
	monitorEnabled := strings.ToLower(os.Getenv("GBNT_MONITOR"))
	if monitorEnabled == "true" || monitorEnabled == "1" {
		go func() {
			// Wait for Docker and API to be fully ready
			time.Sleep(5 * time.Second)
			fmt.Println("\n🛡️  Auto-deploying SRE Monitoring Stack (GBNT_MONITOR=true)...")
			if err := monitor.EnsureNetwork(); err != nil {
				fmt.Printf("❌ SRE Monitor: failed to create network: %v\n", err)
				return
			}
			if err := monitor.WriteConfigs(nil); err != nil {
				fmt.Printf("❌ SRE Monitor: failed to write configs: %v\n", err)
				return
			}
			// Pass Gubernator web credentials for Grafana SSO
			webUser := os.Getenv("GBNT_WEB_USER")
			webPass := os.Getenv("GBNT_WEB_PASSWORD")
			if webUser == "" {
				webUser = "admin"
			}
			if webPass == "" {
				webPass = "admin"
			}
			if err := monitor.DeployManagerStack(webUser, webPass); err != nil {
				fmt.Printf("❌ SRE Monitor: deployment failed: %v\n", err)
				return
			}
			if err := monitor.RegisterInDB(db.GetDB()); err != nil {
				fmt.Printf("⚠️  SRE Monitor: failed to register in DB: %v\n", err)
			}
			aqueducts.GenerateHostsFile()
			fmt.Println("\n✅ SRE Monitoring Stack auto-deployed successfully!")
			fmt.Println("  📈 Grafana: http://localhost:3000  |  🔥 Prometheus: http://localhost:9090  |  📋 Loki: http://localhost:3100")
		}()
	}

	gin.SetMode(gin.ReleaseMode)
	r := gin.Default()

	// API Authentication Middleware
	authMiddleware := func(c *gin.Context) {
		expectedToken := os.Getenv("GBNT_API_TOKEN")
		if expectedToken == "" {
			expectedToken = db.GetAPIToken()
		}
		if c.GetHeader("Authorization") != "Bearer "+expectedToken {
			c.AbortWithStatusJSON(401, gin.H{"error": "Unauthorized API access. Invalid or missing Bearer token."})
			return
		}
		c.Next()
	}

	v1 := r.Group("/v1")
	{
		node := v1.Group("/node", authMiddleware)
		{
			node.GET("/ls", NodeListHandler)
			node.POST("/join", NodeJoinHandler)
			node.POST("/heartbeat", NodeHeartbeatHandler)
			node.GET("/tasks/:node_id", NodeTasksHandler)
			node.POST("/tasks/:task_id/status", UpdateTaskStatusHandler)
			node.GET("/:id", NodeInspectHandler)
			node.POST("/:id/role", NodeRoleHandler)
			node.POST("/:id/availability", NodeAvailabilityHandler)
			node.POST("/:id/reboot", NodeRebootHandler)
			node.POST("/:id/leave", NodeLeaveHandler)
			node.POST("/:id/labels", NodeLabelsHandler)
		}

		cluster := v1.Group("/cluster")
		{
			cluster.GET("/token", ClusterTokenHandler)
			cluster.GET("/info", ClusterInfoHandler)
		}

		stack := v1.Group("/stack", authMiddleware)
		{
			stack.POST("/deploy", StackDeployHandler)
			stack.GET("/ls", StackListHandler)
			stack.GET("/:id/services", StackServicesHandler)
			stack.DELETE("/:id", StackRmHandler)
		}

		service := v1.Group("/service", authMiddleware)
		{
			service.GET("/ls", ServiceListHandler)
			service.GET("/:id/tasks", ServiceTasksHandler)
			service.DELETE("/:id", ServiceRmHandler)
			service.POST("/:id/scale", ServiceScaleHandler)
		}

		task := v1.Group("/task", authMiddleware)
		{
			task.GET("/ls", TaskListHandler)
			task.DELETE("/:id", TaskRmHandler)
		}

		corednsRoute := v1.Group("/coredns", authMiddleware)
		{
			corednsRoute.GET("/config", GetCoreDNSConfig)
			corednsRoute.PUT("/config", UpdateCoreDNSConfig)
		}
	}

	srv := &http.Server{Addr: ":4000", Handler: r}

	go func() {
		<-ctx.Done()
		shutCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
		defer cancel()
		if err := srv.Shutdown(shutCtx); err != nil {
			slog.Error("API server shutdown error", "err", err)
		}
	}()

	slog.Info("starting Gubernator Manager API", "addr", ":4000")
	if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
		return fmt.Errorf("API server: %w", err)
	}
	return nil
}

// startTelemetryServer runs on port 4002 serving Swagger, Metrics and Healthchecks.
func startTelemetryServer(ctx context.Context) {
	r := gin.New()
	r.Use(gin.Recovery())

	r.GET("/health", func(c *gin.Context) {
		c.JSON(200, gin.H{"status": "healthy"})
	})
	r.GET("/metrics", telemetry.MetricsHandler())
	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	srv := &http.Server{Addr: ":4002", Handler: r}

	go func() {
		<-ctx.Done()
		shutCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
		defer cancel()
		if err := srv.Shutdown(shutCtx); err != nil {
			slog.Error("telemetry server shutdown error", "err", err)
		}
	}()

	slog.Info("starting telemetry, health & swagger server", "addr", ":4002")
	if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
		slog.Error("telemetry server error", "err", err)
	}
}

// NodeListResponse represents the response for the node ls endpoint
type NodeListResponse struct {
	Nodes []db.Node `json:"nodes"`
}

// @Summary List all nodes
// @Description Get a list of all nodes currently registered in the Gubernator cluster from SQLite
// @Tags nodes
// @Produce json
// @Success 200 {object} NodeListResponse
// @Router /v1/node/ls [get]
func NodeListHandler(c *gin.Context) {
	var nodes []db.Node
	if err := db.DB.Find(&nodes).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to fetch nodes"})
		return
	}

	c.JSON(http.StatusOK, NodeListResponse{
		Nodes: nodes,
	})
}

// startWatchtowers runs a background loop to update telemetry metrics
// and check for stale nodes (missed heartbeats). Exits when ctx is cancelled.
func startWatchtowers(ctx context.Context) {
	for {
		select {
		case <-ctx.Done():
			return
		case <-time.After(15 * time.Second):
		}

		var nodeCount int64
		db.DB.Model(&db.Node{}).Count(&nodeCount)
		telemetry.TotalNodes.Set(float64(nodeCount))

		var taskCount int64
		db.DB.Model(&db.Task{}).Count(&taskCount)
		telemetry.TotalTasks.Set(float64(taskCount))

		// Mark nodes as 'down' if no heartbeat in 30 seconds
		threshold := time.Now().Add(-30 * time.Second)
		db.DB.Model(&db.Node{}).
			Where("role = ?", "worker").
			Where("updated_at < ?", threshold).
			Where("status = ?", "active").
			Updates(map[string]interface{}{"status": "down"})

		// Also mark tasks of down nodes as dead
		var downNodes []db.Node
		db.DB.Where("role = ? AND status = ?", "worker", "down").Find(&downNodes)
		for _, dn := range downNodes {
			db.DB.Model(&db.Task{}).Where("node_id = ? AND service_id LIKE ?", dn.ID, "core-svc-%").Updates(map[string]interface{}{"status": "dead"})
		}
	}
}